pyproject.toml
[build-system]
requires = ["setuptools>=59.0", "numpy>=1.21.2"]
build-backend = "setuptools.build_meta"

[project]
name = "crimm"
version = "2024.7a1"
authors = [
  { name="Truman Xu", email="ziqiaoxu@umich.edu" },
]
description = "Chemistry with the ReInvented Macromolecular"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy >= 1.21.2",
    "biopython >= 1.80",
    "ipywidgets>=8",
    "nglview>=3.1.2",
    "scipy>=1.7.1",
    "requests>=2.26.0",
    "propka==3.5.0",
    "pandas",
    "rdkit"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Development Status :: 2 - Pre-Alpha",
]

[project.urls]
"Homepage" = "https://github.com/Truman-Xu/crimm"
"Bug Tracker" = "https://github.com/Truman-Xu/crimm/issues"

[tool.setuptools.packages.find]
where = ["./"]
exclude = [
    "src", 
    "__pycache__", 
    ".ipynb_checkpoints", 
    "*.egg-info", 
    "tests"
]
namespaces = false